fix(agent_tools): 修复 bash 工具 timeout float64 转换精度问题
timeout 从 LLM 入参来的是 float64 秒数（如 0.2），原代码
time.Duration(timeout)*time.Second 会先截断后乘，导致 0.2→0 纳秒。
改为 time.Duration(timeout*float64(time.Second)) 先乘后截断，
确保精度正确（0.2*1e9=2e8 纳秒=200ms）。

验证：TestNewBashTool_TimeoutParam 现已通过。

+++ agent_tools/bash.go
@@ -99,7 +99,7 @@ func NewBashTool() *Tools {
 				dangerouslyDisableSandbox = false
 			}
 			//执行工具
-			output, err := bashTool(command, description, time.Duration(timeout)*time.Second, dangerouslyDisableSandbox)
+			output, err := bashTool(command, description, time.Duration(timeout*float64(time.Second)), dangerouslyDisableSandbox)
 			if err != nil {
 				return jsonErr(err.Error())
 			}
